requirements.txt
dash
dash-core-components
dash-html-components
dash-table
Flask
plotly
werkzeug
zmq
PYaml
dash_bootstrap_components